try to improve performance with use of type 1 byte int

+++ ising.h
@@ -1,5 +1,5 @@
 #include "params.h"
 
-void update(const float temp, int grid[L][L]);
+void update(const float temp, int8_t grid[L][L]);
 
-double calculate(int grid[L][L], int* M_max);
+double calculate(int8_t grid[L][L], int* M_max);
